+++ example/app/views.py
@@ -2,7 +2,9 @@
 # coding: utf-8
 
 from django.shortcuts import render
-from tables import PersonTable, ScheduleTable
+from tables import (PersonTable, ScheduleTable, SequenceColumnTable,
+                    AjaxDataTable)
+
 
 class Foo(object):
     def __init__(self, id, name, calendar):
@@ -15,10 +17,23 @@ def base(request):
     people = PersonTable()
     return render(request, "index.html", {'people': people})
 
+
 def edit(request, id):
     pass
+
+
+def sequence_column(request):
+    data = [Foo(1, 'A', [1,2,3,4,5]), Foo(2, 'B', [1,2,3,4,5]), Foo(3, 'C', [1,2,3,4,5])]
+    table = SequenceColumnTable(data)
+    return render(request, "index.html", {'people': table})
+
 
 def calendar_column(request):
     data = [Foo(1, 'A', [1,2,3,4,5,6]), Foo(2, 'B', [1,2,3,4,5,6]), Foo(3, 'C', [1,2,3,4,5,6])]
     table = ScheduleTable(data)
+    return render(request, "index.html", {'people': table})
+
+
+def ajax_data(request):
+    table = AjaxDataTable()
     return render(request, "index.html", {'people': table})