+++ table/__init__.py
@@ -2,4 +2,4 @@
 # coding: utf-8
 
 
-from tables import Table
+from tables import Table, TableData