+++ table/columns/__init__.py
@@ -2,7 +2,8 @@
 # coding: utf-8
 
 from .base import Column, BoundColumn
-from .linkcolumn import LinkColumn, Link
+from .linkcolumn import LinkColumn, Link, ImageLink
 from .datetimecolumn import DatetimeColumn
 from .calendarcolumn import MonthsColumn, WeeksColumn, DaysColumn
 from .sequencecolumn import SequenceColumn
+from .checkboxcolumn import CheckboxColumn