+++ table/columns/base.py
@@ -9,13 +9,13 @@ from django.utils.safestring import mark_safe
 class Column(object):
     """ Represents a single column.
     """
-    
+
     instance_order = 0
 
     def __init__(self, field=None, header=None, attrs=None, header_attrs=None,
                  header_row_order=0, sortable=True, searchable=True, safe=True,
                  visible=True, space=True):
-        self.accessor = Accessor(field)
+        self.field = field
         self.attrs = attrs or {}
         self.sortable = sortable
         self.searchable = searchable
@@ -31,10 +31,10 @@ class Column(object):
         return self.header.text
 
     def render(self, obj):
-        return self.accessor.resolve(obj)
+        return Accessor(self.field).resolve(obj)
 
 class BoundColumn(object):
-    """ A run-time version of Column. The difference between 
+    """ A run-time version of Column. The difference between
         BoundColumn and Column is that BoundColumn objects include the
         relationship between a Column and a object. In practice, this
         means that a BoundColumn knows the "field value" given to the
@@ -43,9 +43,10 @@ class BoundColumn(object):
     def __init__(self, obj, column):
         self.obj = obj
         self.column = column
-        self.base_attrs = column.attrs
-        
+        self.base_attrs = column.attrs.copy()
+
         # copy non-object-related attributes to self directly
+        self.field = column.field
         self.sortable = column.sortable
         self.searchable = column.searchable
         self.safe = column.safe
@@ -54,15 +55,20 @@ class BoundColumn(object):
 
     @property
     def html(self):
-        return self.column.render(self.obj) or ''
+        text = self.column.render(self.obj)
+        if text is None:
+            return ''
+        else:
+            return text
 
     @property
     def attrs(self):
         attrs = {}
-        context = self.obj
         for attr_name, attr in self.base_attrs.items():
-            if isinstance(attr, Accessor):
-                attrs[attr_name] = attr.resolve(context)
+            if callable(attr):
+                attrs[attr_name] = attr(self.obj, self.field)
+            elif isinstance(attr, Accessor):
+                attrs[attr_name] = attr.resolve(self.obj)
             else:
                 attrs[attr_name] = attr
         return mark_safe(' '.join(['%s="%s"' % (attr_name, attr) for attr_name, attr in attrs.items()]))