+++ table/tests/test_views.py
@@ -1,0 +1,140 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+from uuid import uuid4
+from django.test import Client, TestCase
+from django.core.urlresolvers import reverse
+from django.utils import simplejson as json
+from table.views import FeedDataView
+from table.forms import QueryDataForm
+from table.models import Person
+from table.columns import Column
+from table import Table
+
+
+class TestTable(Table):
+    id = Column('id', header='#')
+    name = Column('name', header='NAME')
+    email = Column('email', header='EMAIL', searchable=False)
+
+    class Meta:
+        model = Person
+        ajax = True
+
+
+class FeedDataViewTestCase(TestCase):
+    def setUp(self):
+        self.table = TestTable()
+        self.client = Client()
+        self.url = reverse("feed_data", args=(self.table.token,))
+        self.payload = {
+            "sEcho": "1",
+            "iColumns": 2,
+            "iDisplayStart": 0,
+            "iDisplayLength": 10,
+            "sSearch": "",
+            "bRegex": False,
+            "iSortingCols": 1
+        }
+
+        Person.objects.create(id=1, name="Tom", email="tom@mail.com")
+        Person.objects.create(id=2, name="Jerry", email="jerry@mail.com")
+
+    def test_basic(self):
+        response = self.client.get(self.url, self.payload)
+        self.assertEqual(response.status_code, 200)
+
+        data = json.loads(response.content)
+        expect_data = {
+            "sEcho": "1",
+            "iTotalRecords": 2,
+            "iTotalDisplayRecords": 2,
+            "aaData": [[1, "Tom", "tom@mail.com"], [2, "Jerry", "jerry@mail.com"]]
+            }
+        self.assertEqual(data, expect_data)
+
+    def test_search(self):
+        url, payload = self.url, self.payload
+        payload.update({"sSearch": "T"})
+
+        response = self.client.get(url, payload)
+        self.assertEqual(response.status_code, 200)
+
+        data = json.loads(response.content)
+        expect_data = {
+            "sEcho": "1",
+            "iTotalRecords": 2,
+            "iTotalDisplayRecords": 1,
+            "aaData": [[1, "Tom", "tom@mail.com"]]
+        }
+        self.assertEqual(data, expect_data)
+
+    def test_search_fuzzy(self):
+        url, payload = self.url, self.payload
+        payload.update({"sSearch": "T 2"})
+
+        response = self.client.get(url, payload)
+        self.assertEqual(response.status_code, 200)
+
+        data = json.loads(response.content)
+        expect_data = {
+            "sEcho": "1",
+            "iTotalRecords": 2,
+            "iTotalDisplayRecords": 0,
+            "aaData": []
+        }
+        self.assertEqual(data, expect_data)
+        
+    def test_unsearchable_column(self):
+        url, payload = self.url, self.payload
+        payload.update({"sSearch": "mail"})
+
+        response = self.client.get(url, payload)
+        self.assertEqual(response.status_code, 200)
+
+        data = json.loads(response.content)
+        expect_data = {
+            "sEcho": "1",
+            "iTotalRecords": 2,
+            "iTotalDisplayRecords": 0,
+            "aaData": []
+        }
+        self.assertEqual(data, expect_data)
+
+    def test_sort_asc(self):
+        url, payload = self.url, self.payload
+        payload.update({
+            "iSortCol_0": 0,
+            "sSortDir_0": "asc",
+        })
+
+        response = self.client.get(url, payload)
+        self.assertEqual(response.status_code, 200)
+
+        data = json.loads(response.content)
+        expect_data = {
+            "sEcho": "1",
+            "iTotalRecords": 2,
+            "iTotalDisplayRecords": 2,
+            "aaData": [[1, "Tom", "tom@mail.com"], [2, "Jerry", "jerry@mail.com"]]
+        }
+        self.assertEqual(data, expect_data)
+
+    def test_sort_desc(self):
+        url, payload = self.url, self.payload
+        payload.update({
+            "iSortCol_0": 0,
+            "sSortDir_0": "desc",
+        })
+
+        response = self.client.get(url, payload)
+        self.assertEqual(response.status_code, 200)
+
+        data = json.loads(response.content)
+        expect_data = {
+            "sEcho": "1",
+            "iTotalRecords": 2,
+            "iTotalDisplayRecords": 2,
+            "aaData": [[2, "Jerry", "jerry@mail.com"], [1, "Tom", "tom@mail.com"]]
+        }
+        self.assertEqual(data, expect_data)